Replace backend with advanced Flask implementation
