--- backend/package.json
@@ -1,41 +1,0 @@
-
-{
-  "name": "bootstrap-vs-zombies-backend",
-  "version": "1.0.0",
-  "description": "Backend API for Bootstrap vs Zombies educational game",
-  "main": "dist/server.js",
-  "scripts": {
-    "dev": "nodemon src/server.ts",
-    "build": "tsc",
-    "start": "node dist/server.js",
-    "test": "jest"
-  },
-  "keywords": ["bootstrap", "education", "gaming", "api"],
-  "author": "Bootstrap vs Zombies Team",
-  "license": "MIT",
-  "dependencies": {
-    "express": "^4.18.2",
-    "cors": "^2.8.5",
-    "helmet": "^7.1.0",
-    "bcryptjs": "^2.4.3",
-    "jsonwebtoken": "^9.0.2",
-    "pg": "^8.11.3",
-    "dotenv": "^16.3.1",
-    "express-rate-limit": "^7.1.5",
-    "express-validator": "^7.0.1"
-    , "nodemailer": "^6.9.11"
-  },
-  "devDependencies": {
-    "@types/express": "^4.17.21",
-    "@types/cors": "^2.8.17",
-    "@types/bcryptjs": "^2.4.6",
-    "@types/jsonwebtoken": "^9.0.5",
-    "@types/pg": "^8.10.9",
-    "@types/node": "^20.10.5",
-    "typescript": "^5.3.3",
-    "nodemon": "^3.0.2",
-    "ts-node": "^10.9.2",
-    "jest": "^29.7.0",
-    "@types/jest": "^29.5.8"
-  }
-}
