+++ backend/requirements.txt
@@ -1,2 +1,6 @@
 Flask==3.0.3
+Flask-JWT-Extended
+Flask-Bcrypt
+Flask-SQLAlchemy
+python-dotenv
 gunicorn==21.2.0
