--- backend/src/routes/auth.ts
@@ -1,282 +1,0 @@
-
-/**
- * Authentication Routes
- * 
- * Handles user registration, login, logout, and token management.
- * Implements JWT-based authentication with secure password hashing.
- * 
- * Educational Value:
- * - Demonstrates secure authentication patterns
- * - Shows JWT token implementation
- * - Implements password hashing best practices
- * - Provides user session management
- */
-
-import express from 'express';
-import bcrypt from 'bcryptjs';
-import jwt from 'jsonwebtoken';
-import crypto from 'crypto';
-import { body, validationResult } from 'express-validator';
-import { authenticateToken } from '../middleware/auth';
-import { sendPasswordResetEmail } from '../utils/email';
-
-const router = express.Router();
-
-/**
- * Temporary in-memory user storage
- * In production, this would be replaced with a database
- */
-interface User {
-  id: string;
-  email: string;
-  name: string;
-  password: string;
-  createdAt: Date;
-  gameStats: {
-    totalGames: number;
-    highScore: number;
-    totalScore: number;
-  };
-  resetToken?: string;
-  resetTokenExpiry?: Date;
-}
-
-const users: User[] = [];
-
-/**
- * POST /api/auth/register
- * Register a new user account
- */
-router.post('/register', [
-  body('email').isEmail().normalizeEmail(),
-  body('name').trim().isLength({ min: 2, max: 50 }),
-  body('password').isLength({ min: 6 }).matches(/^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)/),
-], async (req, res) => {
-  try {
-    // Validate input
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(400).json({
-        error: 'Validation failed',
-        details: errors.array()
-      });
-    }
-
-    const { email, name, password } = req.body;
-
-    // Check if user already exists
-    const existingUser = users.find(user => user.email === email);
-    if (existingUser) {
-      return res.status(409).json({
-        error: 'User already exists',
-        message: 'An account with this email already exists'
-      });
-    }
-
-    // Hash password
-    const saltRounds = 12;
-    const hashedPassword = await bcrypt.hash(password, saltRounds);
-
-    // Create new user
-    const newUser: User = {
-      id: Date.now().toString(),
-      email,
-      name,
-      password: hashedPassword,
-      createdAt: new Date(),
-      gameStats: {
-        totalGames: 0,
-        highScore: 0,
-        totalScore: 0
-      }
-    };
-
-    users.push(newUser);
-
-    // Generate JWT token
-    const token = jwt.sign(
-      { userId: newUser.id, email: newUser.email },
-      process.env.JWT_SECRET || 'fallback-secret-key',
-      { expiresIn: '7d' }
-    );
-
-    // Return user data (without password)
-    const { password: _, ...userWithoutPassword } = newUser;
-    
-    res.status(201).json({
-      message: 'User registered successfully',
-      user: userWithoutPassword,
-      token
-    });
-
-  } catch (error) {
-    console.error('Registration error:', error);
-    res.status(500).json({
-      error: 'Internal server error',
-      message: 'Failed to register user'
-    });
-  }
-});
-
-/**
- * POST /api/auth/login
- * Authenticate user and return JWT token
- */
-router.post('/login', [
-  body('email').isEmail().normalizeEmail(),
-  body('password').exists()
-], async (req, res) => {
-  try {
-    // Validate input
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(400).json({
-        error: 'Validation failed',
-        details: errors.array()
-      });
-    }
-
-    const { email, password } = req.body;
-
-    // Find user
-    const user = users.find(u => u.email === email);
-    if (!user) {
-      return res.status(401).json({
-        error: 'Invalid credentials',
-        message: 'Email or password is incorrect'
-      });
-    }
-
-    // Verify password
-    const isValidPassword = await bcrypt.compare(password, user.password);
-    if (!isValidPassword) {
-      return res.status(401).json({
-        error: 'Invalid credentials',
-        message: 'Email or password is incorrect'
-      });
-    }
-
-    // Generate JWT token
-    const token = jwt.sign(
-      { userId: user.id, email: user.email },
-      process.env.JWT_SECRET || 'fallback-secret-key',
-      { expiresIn: '7d' }
-    );
-
-    // Return user data (without password)
-    const { password: _, ...userWithoutPassword } = user;
-    
-    res.json({
-      message: 'Login successful',
-      user: userWithoutPassword,
-      token
-    });
-
-  } catch (error) {
-    console.error('Login error:', error);
-    res.status(500).json({
-      error: 'Internal server error',
-      message: 'Failed to authenticate user'
-    });
-  }
-});
-
-/**
- * GET /api/auth/me
- * Get current user profile (requires authentication)
- */
-router.get('/me', authenticateToken, (req, res) => {
-  try {
-    const userId = (req as any).user?.userId;
-    const user = users.find(u => u.id === userId);
-
-    if (!user) {
-      return res.status(404).json({
-        error: 'User not found',
-        message: 'User account no longer exists'
-      });
-    }
-
-    // Return user data (without password)
-    const { password: _, ...userWithoutPassword } = user;
-    
-    res.json({
-      user: userWithoutPassword
-    });
-
-  } catch (error) {
-    console.error('Profile fetch error:', error);
-    res.status(500).json({
-      error: 'Internal server error',
-      message: 'Failed to fetch user profile'
-    });
-  }
-});
-
-/**
- * POST /api/auth/forgot-password
- * Initiate password reset process
- */
-router.post('/forgot-password', [body('email').isEmail().normalizeEmail()], async (req, res) => {
-  const errors = validationResult(req);
-  if (!errors.isEmpty()) {
-    return res.status(400).json({ error: 'Validation failed', details: errors.array() });
-  }
-
-  const { email } = req.body;
-  const user = users.find(u => u.email === email);
-
-  if (user) {
-    const token = crypto.randomBytes(32).toString('hex');
-    user.resetToken = token;
-    user.resetTokenExpiry = new Date(Date.now() + 60 * 60 * 1000); // 1 hour
-    try {
-      await sendPasswordResetEmail(email, token);
-    } catch (err) {
-      console.error('Failed to send reset email:', err);
-    }
-  }
-
-  res.json({ message: 'If the email exists, a reset link has been sent.' });
-});
-
-/**
- * POST /api/auth/reset-password
- * Complete password reset
- */
-router.post('/reset-password', [
-  body('token').exists(),
-  body('password').isLength({ min: 6 })
-], async (req, res) => {
-  const errors = validationResult(req);
-  if (!errors.isEmpty()) {
-    return res.status(400).json({ error: 'Validation failed', details: errors.array() });
-  }
-
-  const { token, password } = req.body;
-  const user = users.find(u => u.resetToken === token && u.resetTokenExpiry && u.resetTokenExpiry > new Date());
-
-  if (!user) {
-    return res.status(400).json({ error: 'Invalid or expired token' });
-  }
-
-  const hashedPassword = await bcrypt.hash(password, 12);
-  user.password = hashedPassword;
-  user.resetToken = undefined;
-  user.resetTokenExpiry = undefined;
-
-  res.json({ message: 'Password reset successful' });
-});
-
-/**
- * POST /api/auth/logout
- * Logout user (in a stateless JWT system, this is mainly for client-side cleanup)
- */
-router.post('/logout', authenticateToken, (req, res) => {
-  res.json({
-    message: 'Logout successful',
-    note: 'Please remove the token from client storage'
-  });
-});
-
-export default router;
