--- backend/src/utils/email.ts
@@ -1,24 +1,0 @@
-import nodemailer from 'nodemailer';
-
-export const sendPasswordResetEmail = async (to: string, token: string) => {
-  const transporter = nodemailer.createTransport({
-    host: process.env.SMTP_HOST,
-    port: parseInt(process.env.SMTP_PORT || '587', 10),
-    secure: false,
-    auth: {
-      user: process.env.SMTP_USER,
-      pass: process.env.SMTP_PASS
-    }
-  });
-
-  const resetUrl = `${process.env.FRONTEND_URL || 'http://localhost:5173'}/reset-password?token=${token}`;
-
-  await transporter.sendMail({
-    from: process.env.SMTP_FROM || 'no-reply@example.com',
-    to,
-    subject: 'Password Reset',
-    text: `Reset your password: ${resetUrl}`,
-    html: `<p>Click <a href="${resetUrl}">here</a> to reset your password.</p>`
-  });
-};
-
