+++ docker-compose.yml
@@ -14,7 +14,8 @@ services:
       - backend
 
   backend:
-    build: ./backend
+    build:
+      context: ./backend
     ports:
       - "5000:5000"
     environment:
